refactoring some things
