+++ sizebayfront/src/pages/Home/index.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-mixed-operators */
 import React, { useContext, useState, useEffect } from 'react';
 
-import { itemContext } from '../../contexts/ItemContext';
+import { ItemContext } from '../../contexts/item';
 
 import ItemAddBar from '../../components/ItemAddBar';
 
@@ -9,7 +9,7 @@ import Item from '../../components/Item';
 import { ItemsArea } from '../../components/Item/styles';
 
 export default function Home() {
-  const { itemsCollection, isSearch, searchText } = useContext(itemContext);
+  const { itemsCollection, isSearch, searchText } = useContext(ItemContext);
   const [foundedItems, setFoundedItems] = useState([]);
 
   function searchItem() {
