+++ sizebayfront/src/pages/Pending/index.js
@@ -2,13 +2,13 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import React, { useContext, useState, useEffect } from 'react';
 
-import { itemContext } from '../../contexts/ItemContext';
+import { ItemContext } from '../../contexts/item';
 
 import Item from '../../components/Item';
 import { ItemsArea } from '../../components/Item/styles';
 
 export default function Pending() {
-  const { itemsCollection, isSearch, searchText } = useContext(itemContext);
+  const { itemsCollection, isSearch, searchText } = useContext(ItemContext);
   const [foundedItems, setFoundedItems] = useState([]);
 
   const filteredCollection = itemsCollection.filter(item => item.isPending === true);
